requirements/base.txt
alabaster==0.7.9
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
arrow==0.10.0
asn1crypto==0.24.0
awesome-slugify==1.6.5
Babel==2.3.4
billiard==3.3.0.23
binaryornot==0.4.0
blockchain==1.3.3
celery==3.1.24
certifi==2018.1.18
cffi==1.11.5
chardet==2.3.0
click==6.7
cloudinary==1.11.0
cookiecutter==1.5.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.3.4
cryptography==2.2.2
decorator==4.0.11
defusedxml==0.5.0
Django==1.11
django-allauth==0.30.0
django-braces==1.11.0
django-cors-headers==2.0.2
django-coverage-plugin==1.4.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-environ==0.4.1
django-extensions==1.7.6
django-formtools==2.1
django-model-utils==2.6.1
django-otp==0.4.3
django-phonenumber-field==1.3.0
django-redis==4.7.0
django-rest-auth==0.9.1
django-rest-swagger==2.2.0
django-test-plus==1.0.17
django-two-factor-auth==1.7.0
djangoblockchain-info==0.1.0
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
docutils==0.13.1
drfdocs==0.0.11
factory-boy==2.8.1
Faker==0.7.7
future==0.16.0
idna==2.6
imagesize==0.7.1
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.5
jinja2-time==0.2.0
kombu==3.0.37
MarkupSafe==0.23
mock==2.0.0
nexmo==2.0.0
oauthlib==2.0.1
olefile==0.44
openapi-codec==1.3.2
pbr==4.0.0
pexpect==4.2.1
phonenumberslite==8.9.3
pickleshare==0.7.4
Pillow==4.0.0
poyo==0.4.0
prompt-toolkit==1.0.13
psycopg2==2.7.3.1
ptyprocess==0.5.1
py==1.4.32
pycparser==2.18
Pygments==2.2.0
PyJWT==1.4.2
pytest==3.0.6
pytest-django==3.1.2
pytest-sugar==0.8.0
python-dateutil==2.6.0
python-http-client==3.0.0
python3-openid==3.0.10
pytz==2016.10
qrcode==4.0.4
redis==2.10.5
regex==2017.2.8
requests==2.13.0
requests-oauthlib==0.7.0
sendgrid==3.6.5
sendgrid-django==4.0.4
simplegeneric==0.8.1
simplejson==3.14.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.2
sqlparse==0.2.2
termcolor==1.1.0
traitlets==4.3.1
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.11.15
whichcraft==0.4.0
whitenoise==3.3.0
